// frontend/src/shared/api/generated/models/chatMessage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Databricks Apps FastAPI Starter API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessagePartsItem } from './chatMessagePartsItem';

export interface ChatMessage {
  content: string;
  createdAt: string;
  id: string;
  parts: ChatMessagePartsItem[];
  role: string;
  traceId?: string | null;
}